+++ ext/tree_sitter/node.c
@@ -165,15 +165,14 @@ static VALUE node_child_by_field_id(VALUE self, VALUE field_id) {
  * @return [Node]
  */
 static VALUE node_child_by_field_name(VALUE self, VALUE field_name) {
-  if (Qtrue == rb_funcall(self, rb_intern("field?"), 1, field_name)) {
-    VALUE field_str = rb_funcall(field_name, rb_intern("to_s"), 0);
-    const char *name = StringValuePtr(field_str);
-    uint32_t length = (uint32_t)RSTRING_LEN(field_str);
-    return new_node_by_val(ts_node_child_by_field_name(SELF, name, length),
-                           unwrap(self)->tree);
-  } else {
+  VALUE field_str = rb_funcall(field_name, rb_intern("to_s"), 0);
+  const char *name = StringValuePtr(field_str);
+  uint32_t length = (uint32_t)RSTRING_LEN(field_str);
+  TSNode child = ts_node_child_by_field_name(SELF, name, length);
+  if (ts_node_is_null(child)) {
     return Qnil;
   }
+  return new_node_by_val(child, unwrap(self)->tree);
 }
 
 /**